Factor file io into its own module
