+++ Dark Samurai Warrior Reloaded/io/file.h
@@ -1,0 +1,9 @@
+#pragma once
+#include <stddef.h>
+
+typedef struct LoadedFile {
+  size_t size;
+  void *memory;
+} LoadedFile;
+
+LoadedFile win32_load_file(char *filename);
