+++ Dark Samurai Warrior Reloaded/main.c
@@ -9,6 +9,7 @@
 #include "input/input.h"
 #include "math.h"
 #include "render.h"
+#include "io/file.h"
 
 typedef struct Win32Buffer {
   BITMAPINFO info;
@@ -42,10 +43,6 @@ static bool global_running;
 // https://learn.microsoft.com/en-us/windows/win32/api/profileapi/nf-profileapi-queryperformancefrequency
 static u32 performance_frequency;
 
-typedef struct LoadedFile {
-  size_t size;
-  void *memory;
-} LoadedFile;
 
 #pragma pack(push, 1)
 typedef struct BitmapHeader {
@@ -87,24 +84,6 @@ inline LARGE_INTEGER win32_get_wall_clock() {
 inline float win32_get_seconds_elapsed(LARGE_INTEGER start, LARGE_INTEGER end) {
   float result =
       ((float)(end.QuadPart - start.QuadPart) / (float)performance_frequency);
-  return result;
-}
-
-LoadedFile win32_load_file(char *filename) {
-  LoadedFile result = {0};
-  HANDLE file_handle = CreateFileA(filename, GENERIC_READ, FILE_SHARE_READ, 0,
-                                   OPEN_EXISTING, FILE_ATTRIBUTE_NORMAL, 0);
-  assert(file_handle != INVALID_HANDLE_VALUE);
-  LARGE_INTEGER file_size64;
-  assert(GetFileSizeEx(file_handle, &file_size64) != INVALID_FILE_SIZE);
-  size_t file_size32 = file_size64.QuadPart;
-  result.memory =
-      VirtualAlloc(0, file_size32, MEM_RESERVE | MEM_COMMIT, PAGE_READWRITE);
-  DWORD bytes_read;
-  assert(ReadFile(file_handle, result.memory, file_size32, &bytes_read, 0));
-  assert(bytes_read == file_size32);
-  result.size = file_size32;
-
   return result;
 }
